# Improved performance using numpy sorted union and intersect