Hi there, I was interested in your implementation and got curious if its performance could be improved by leveraging numpy arrays directly and doing sorted unions/intersects instead of relying on pyroaring's BitMaps.

Keeping initialisation overhead down to match the original implementation ended up being the most difficult, but I got pretty close. Init overhead comparison from the benchmark script can be found below:

```
...with 100 rows and 10 loops, cube init in 0.00081 sec, cube2 init in 0.00089 sec, overall in 0.02116 sec.
...with 200 rows and 10 loops, cube init in 0.00061 sec, cube2 init in 0.00124 sec, overall in 0.02050 sec.
...with 400 rows and 10 loops, cube init in 0.00118 sec, cube2 init in 0.00199 sec, overall in 0.02318 sec.
...with 800 rows and 10 loops, cube init in 0.00172 sec, cube2 init in 0.00344 sec, overall in 0.02799 sec.
...with 1,600 rows and 10 loops, cube init in 0.00290 sec, cube2 init in 0.00539 sec, overall in 0.03604 sec.
...with 3,200 rows and 10 loops, cube init in 0.00519 sec, cube2 init in 0.00886 sec, overall in 0.05205 sec.
...with 6,400 rows and 10 loops, cube init in 0.00907 sec, cube2 init in 0.01470 sec, overall in 0.08270 sec.
...with 12,800 rows and 10 loops, cube init in 0.01587 sec, cube2 init in 0.02405 sec, overall in 0.14035 sec.
...with 25,600 rows and 10 loops, cube init in 0.02788 sec, cube2 init in 0.03808 sec, overall in 0.24720 sec.
...with 51,200 rows and 10 loops, cube init in 0.05071 sec, cube2 init in 0.06107 sec, overall in 0.45752 sec.
...with 102,400 rows and 10 loops, cube init in 0.09740 sec, cube2 init in 0.10827 sec, overall in 0.86423 sec.
...with 204,800 rows and 10 loops, cube init in 0.18905 sec, cube2 init in 0.19851 sec, overall in 1.70267 sec.
...with 409,600 rows and 10 loops, cube init in 0.37383 sec, cube2 init in 0.38502 sec, overall in 3.37266 sec.
...with 819,200 rows and 10 loops, cube init in 0.75209 sec, cube2 init in 0.76295 sec, overall in 6.84906 sec.
...with 1,638,400 rows and 10 loops, cube init in 1.47288 sec, cube2 init in 1.49330 sec, overall in 13.50181 sec.
...with 3,276,800 rows and 10 loops, cube init in 2.93813 sec, cube2 init in 3.01520 sec, overall in 27.32047 sec.
...with 6,553,600 rows and 10 loops, cube init in 5.85592 sec, cube2 init in 5.94174 sec, overall in 54.03826 sec.
...with 13,107,200 rows and 10 loops, cube init in 11.72648 sec, cube2 init in 11.96820 sec, overall in 108.82562 sec.
```

And here are all the graphs produced by the benchmark script, the alternative implementation is marked as "Cube2"

![hk](https://github.com/user-attachments/assets/1f072479-a01d-4444-b5d2-0e7663f8690e)
![xl](https://github.com/user-attachments/assets/8ea8d0eb-1f1f-446c-9418-6ccd197bc39d)
![l](https://github.com/user-attachments/assets/616903ca-2c98-42e5-a4a4-a591eee42f64)
![m](https://github.com/user-attachments/assets/d33a7a48-2242-4976-afeb-8dd8d9946f32)
![s](https://github.com/user-attachments/assets/eb741a80-fa42-46c0-a95b-82825c224f44)

As you can see the XL performance now also consistently beats pandas on any dataframe size

This PR is not prepared to be merged as it still includes both implementations to allow anyone to run these comparisons, but rather serves as a side track which can be copied from if you would like to incorporate these performance improvements.