@Zeutschler 2024-10-10
Great contribution!!! As I anyhow need to unravel the codebase, I will isolate the index so we can easily switch between different indexing approaches. Your code will become one engine.