@Zeutschler 2024-10-14
I've detected a small but important bug in my code. The following code returns a Roaring Bitmap object, instead of an (nd)array: `records = reduce(lambda x, y: x & y, bitmaps) ...`
When Numpy now tries to resolve `np.nansum(self.values[i][records])` it likely uses the build in iterator of the Roaring Bitmap implementation which seems to be very expensive. By returning an array instead
'records = reduce(lambda x, y: x & y, bitmaps).**to_array()**' the aggregation performance is very close or equal your (@code-mc) Cube2 implementation. 

I will now compare both implementations (Roaring vs. ndarray based) for memory consumption and serialization speed. If there is no significant difference I will likely switch the ndarray + snp-based approach as it is simpler and comes with other advantages (e.g. random access  for binary search in arrays).