@Zeutschler 2024-10-14
I turned out that using Numpy arrays + sortednp is only faster, if NO intersection operations needs to be performed, but just simple lookups. The better performance is due to the fact, that Roaring Bitmaps require a dedicated deserialisation step. Once one or more intersections need to be executed, Roaring Bitmaps are orders of magnitudes faster, especially at larger datasets. The benchmarks have been updated to include both methods. Again, many thanks to @code-mc for the super valuable input!